Move making a request to get case-correct username before connecting to client/caseCorrect
