+++ src/client/caseCorrect.js
@@ -1,0 +1,43 @@
+var yggdrasil = require('yggdrasil')({});
+var UUID = require('uuid-1345');
+
+module.exports = function(client, options) {
+  var clientToken = options.clientToken || UUID.v4().toString();
+  var accessToken;
+  var haveCredentials = options.password != null || (clientToken != null && options.session != null);
+
+  if(haveCredentials) {
+    // make a request to get the case-correct username before connecting.
+    var cb = function(err, session) {
+      if(err) {
+        client.emit('error', err);
+      } else {
+        client.session = session;
+        client.username = session.selectedProfile.name;
+        accessToken = session.accessToken;
+        client.emit('session');
+        client.connect(options.port, options.host);
+      }
+    };
+
+    if (options.session) {
+      yggdrasil.validate(options.session.accessToken, function(ok) {
+        if (ok)
+          cb(null, options.session);
+        else
+          yggdrasil.refresh(options.session.accessToken, options.session.clientToken, function(err, _, data) {
+            cb(err, data);
+          });
+      });
+    }
+    else yggdrasil.auth({
+      user: options.username,
+      pass: options.password,
+      token: clientToken
+    }, cb);
+  } else {
+    // assume the server is in offline mode and just go for it.
+    client.username = options.username;
+    client.connect(options.port, options.host);
+  }
+};
